fix(user-domain): rename Entity folder to Entities
