+++ src/User/Repository/UserRepository.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\User\Repository;
 
-use App\User\Entity\User;
+use App\User\Entities\User;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
